final commit

+++ app/Http/Controllers/TaskController.php
@@ -6,6 +6,9 @@ use Illuminate\Http\Request;
 use App\Task;
 use App\Http\Requests\TaskRequest;
 use Auth;
+use Response;
+use SoapBox\Formatter\Formatter;
+
 class TaskController extends Controller
 {
     // Change the state of a given task
@@ -25,12 +28,13 @@ class TaskController extends Controller
     // Open a form to edit some task
     public function edit($id){
         $task = Task::find($id);                                    // Get the task
+        if($task == null)
+            return view('msg_only')->with(['danger' => "This task doesn't exist."]); // If the task Doesn't exist
         return view('tasks.edit_form')->with(['task' => $task]);    // Return task info to a view form
     }
 
     // Update some task
     public function update(TaskRequest $request){
-
         $result = $request->only(['id', 'name', 'description']);    // Get the request parameters
         $task = Task::find($result['id']);                          // Find the test with this id
         $task->name = $result['name'];                              // Change the informations
@@ -56,12 +60,35 @@ class TaskController extends Controller
     // Delete some task
     public function delete($id){
         $task = Task::find($id);
-        $msg = "You have no permission to delete this task";
+        if($task == null)
+            return view('msg_only')->with(['danger' => "This task doesn't exist."]); // If the task Doesn't exist
 
         if(Auth::user()->can('delete-task', $task)){                // The user can only delete a task according to
             $task->delete();                                        // the police gate 'delete-task' defined in AuthServiceProvider
-            $msg = "Task sucessfully deleted.";
+            return redirect()->action('HomeController@index')->with(['msg' => "Task sucessfully deleted."]);
         }
-        return redirect()->action('HomeController@index')->with(['msg' => $msg]);
+        return redirect()->action('HomeController@index')->with(['msg' => "You don't have permission to delete this task"]);
+    }
+
+    // Generate CVS table
+    public function cvs_table(){
+        $tasks = Task::all();                                                           // Get all tasks
+        $handle = fopen("cvs_table.csv", 'w+');                                         // Open a file
+        fputcsv($handle, ['Name', 'Description', 'State', 'Created in','Last Update']); // Put the titles in the first line
+        foreach($tasks as $task) {
+            fputcsv($handle, [  $task['name'], $task['description'], $task['state'],    // Write the tasks
+                                $task['created_at'], $task['updated_at']]);
+        }
+        fclose($handle);                                                                // Close the file
+
+        return Response::download("cvs_table.csv", 'cvs_table.csv', ['Content-Type' => 'text/csv']); // Export to the browser
+
+    }
+
+    // Generate XML table
+    public function xml_table(){
+        $formatter = Formatter::make(Task::all(), Formatter::JSON);             //Formarting the task JSON object
+        return Response::make($formatter->toXml(), '200')                       //Returnin to XML
+        ->header('Content-Type', 'text/xml');
     }
 }
